pyproject.toml
[tool.poetry]
name = "pokedex"
version = "0.1.0"
description = ""
authors = ["Mike <michael.yelverton89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.3"
aiofiles = "^22.1.0"
black = "^22.10.0"
pytest = "^7.1.3"
pillow = "^9.2.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
